fix: Separate PHP 8.2+ specific tests to avoid syntax errors
- Move intersection+union type test to FileUploadFactoryPhp82Test.php
- Add @requires PHP >= 8.2 at class level for proper version checking
- Remove unused imports from main test file
- Fix compatibility for PHP 8.1 where intersection+union syntax causes parse errors

This ensures tests run properly on all supported PHP versions.
